WFE-565 package/libopenblt: add package to BR2 config
Add the OpenBLT package and include in the default defconfig for our
project.

OpenBLT is the library used for communicating with the custom CAN
bootloaders installed on the STM32F072RB and STM32F767VI chips included
in the various subsustems of the vehicle.

This library has Python bindings and will be used by a Python script
added in the beaglebone application package.

Select BR2_PACKAGE_LIBOPENBLT to enable.

Also, intializes adding project packages to our BR2_EXTERNAL tree
structure.