WFE-536 Get SGX working
The current configuration of the Texas Instrument SGX drivers included
with Buildroot 2020.08 was broken. Fortunately, there are some patches
supplied by the community that offer some fixes to get it working.

Firstly, we were not including the correct .dtb in our SD card image for
the Beaglebone Black Wireless. This was causing issues with initializing
the kernel driver for the SGX.

Further detail for the patches can be found in 0001-sgx.patch.