Add post-checkout hook to follow master/develop branches in common-all
This is something I think is generally useful for assisting with keeping
our main branches `master/` and `develop/` in sync. It also spits a
warning if for whatever reason common-all is not initialized properly.

I don't think there are any issues/annoyances this would cause. The only
case where I think this would mess up things is if you were working on
something in common-all but needed to switch to develop or master in an
upper project branch if you weren't already there. Hopefully that's more
of a corner case.