WFE-279 Restore right motor controller code
This code was disabled while the right motor controller was not working,
but since we've now corrected issues on the motor and motor controller
side (incorrectly installed encoder, misconfigured proCAN message
priority) we can restore it.

This has been tested on the vehicle, both motors are now working in
conjunction.