Add debouncing for HV/EM toggle button presses
Since the updated DCU code doesn't debounce, we can get many events sent to
the event queue in a row, depending how bouncy the switch is. This can
overflow the event queue. So we should clean up our inputs by debouncing the
signal. It does this by sampling the input with a period of 5ms until we get a
stable value for 16 samples.