CAN was always disabled due to use of invalid defines
The desired behaviour was that the CAN would disabled on Nucleo builds
because they generally won't have a valid CAN bus connection. (This was
not always the case, as we had a Nucleo HITL for development at some
point).

However, the actual behaviour that was introduced was that the boards
would always be disabled because there were some defines used that were
actually not defined, but the preprocessor doesn't throw a warning, and
just always evaluates the expression.

This change was tested for correct code generation with `gcc <...> -E`